social-syrup/phase3_replicate/requirements.txt
replicate==0.34.4
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
